cif: wq: cast cycle to uint64_t for duration comparison
To fixup compiler warning.

+++ subsys/net/lib/sockets/ldp/ldp_utils_ze.cpp
@@ -133,7 +133,7 @@ void ldp_wq::schedule()
 			duration = 0;
 		}
 
-		if (duration > early_wi->cycle) {
+		if (duration > (uint64_t)early_wi->cycle) {
 			LOG_WRN("work item %d callback time %llu is longer than cycle %d",
 				early_wi->id, duration, early_wi->cycle);
 		}
